Phase 12 — CI on push (lint + typecheck + Postgres/SQLite matrix)
Add .github/workflows/lint.yml running ruff and mypy on every push
and PR. Convert build.yml's sequential Postgres-then-SQLite test
steps into a real strategy matrix so the legs run in parallel and
GitHub surfaces them as distinct status checks for branch protection.

Lint surfaced 454 ruff and 274 mypy errors that had accumulated. Ruff
is now fully clean; mypy is clean on src/ under a pragmatic strict
config that keeps bug-catching rules (None-safety, attribute access,
return values) but drops cosmetic ones (untyped-defs, generic-args,
warn-return-any) — full strict on 192 source files would be days of
work for negligible bug-finding payoff. Tests aren't typechecked
since they lean heavily on mocks; pytest still covers their behaviour.

Notable substantive fixes:
- algorithm builtins now subscript history_arrays["close"] instead
  of .get("close") since producers always populate all 5 OHLCV keys
  — eliminates spurious Optional and the dead None checks
- ML algorithms (xgboost_trend, regime_hmm, torch_base) gained
  assert-not-None narrows on stored model/scaler state
- repository.py model_class is now Any so attribute access through
  the generic CRUD helper typechecks
- ClassVar applied to all registry _adapters/_algorithms dicts
- two RUF006 fire-and-forget asyncio.create_task calls now pinned
  in module-level sets so they survive GC
- composite.py guards against None root_node_id
- migrations/env.py raises if alembic.ini lacks sqlalchemy.url
- alpaca adapters narrow union-typed alpaca-py return types
- backtest engine collapses double-nested if into single guard

567 pytest tests still pass.

+++ src/daytrader/ui/pages/home.py
@@ -9,7 +9,7 @@ from __future__ import annotations
 
 from nicegui import ui
 
-from ..services import count_personas, list_discoveries, list_personas
+from ..services import list_discoveries, list_personas
 from ..shell import page_layout, persona_card, stat_card
 
 
@@ -72,7 +72,7 @@ async def _render_regime_summary_card() -> None:
         from ..services_regime import get_current_regime
 
         snap = await get_current_regime()
-    except Exception:  # noqa: BLE001
+    except Exception:
         return
 
     if snap.status != "ok":
@@ -84,7 +84,7 @@ async def _render_regime_summary_card() -> None:
         "bear": "negative",
         "sideways": "warning",
     }.get(snap.regime, "primary")
-    top_pct = int(round(snap.probabilities.get(snap.regime, 0.0) * 100))
+    top_pct = round(snap.probabilities.get(snap.regime, 0.0) * 100)
     stat_card(
         "Market Regime",
         f"{snap.regime.upper()} {top_pct}%",
@@ -127,7 +127,7 @@ async def _render_recent_discoveries() -> None:
     """Top 5 significant discoveries by lift, most recent first."""
     try:
         rows = await list_discoveries(significant_only=True, limit=5)
-    except Exception as exc:  # noqa: BLE001
+    except Exception as exc:
         ui.label(f"Could not load discoveries: {exc}").classes("text-negative text-caption")
         return
     if not rows:
